Add a "PR is up-to-date" check (#19270)
GitHub's built-in "Require branches to be up to date before merging"
seems a bit too strict, as every time someone merges anything it'll
require a master merge and therefore a re-run of all the CI checks.

On the other hand, it would be advantageous to have a check that
requires branches be up to date to the most recent time we did a release
to avoid the need for PRs like #19266.

So what we do here is define a tag indicating the master commit that PRs
need to be up to date with respect to. A push to master runs a job that
might update this tag, and any push to the tag will check all open PRs
and set a status check to "failed" for any that aren't up to date.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/700979826