pr-is-up-to-date: Speed up processing of tag pushes with paths (#23123)
When processing a tag push and PATHS is set, we need to find the merge
base to get the diff to determine if the paths are touched.

But it's a lot faster to fetch 100 revisions at once than to fetch 10
revisions 10 times. So instead of fetching to each PR's merge base,
which will likely enough fetch a few revisions for every PR beyond a
certain age, fetch arbitrary blocks of revisions.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1900302702